move source data
